requirements.txt
py-xbrl
requests
